MixedUnAttemptedQuiz
Known potential bug: The attempted questions are saved and counted up in the same row which mean I can only see whether the last attempt was succsfful or failed. which is good for games suchs as quizes of wrong answered questions until they are answered succsffuly, but impossible to get a full record for quizzes such as the hardest questions for all users or questions particual users struggled with the most.